# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kaynatt API",
        "description": "Mon API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/auth/fairePaiement/{gestionCycle}": {
            "post": {
                "tags": [
                    "Paiement Participant"
                ],
                "summary": "Effectuer un paiement pour un cycle de gestion donné",
                "description": "Effectue un paiement pour un cycle de gestion donné dans le système.",
                "operationId": "effectuerPaiement",
                "parameters": [
                    {
                        "name": "gestionCycle",
                        "in": "path",
                        "description": "ID du cycle de gestion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données requises pour effectuer le paiement",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date_paiement": {
                                        "description": "Date du paiement",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "montant_paiement": {
                                        "description": "Montant du paiement",
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès - Le paiement a été effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "Le paiement a été effectué avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte - Veuillez vérifier les paramètres de la requête"
                    },
                    "401": {
                        "description": "Non autorisé - Veuillez vous connecter pour effectuer cette action"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/createur_tontine/gererCycle/{tontine}": {
            "post": {
                "tags": [
                    "CreateurTontine"
                ],
                "summary": "Gérer les cycles d'une tontine",
                "description": "Crée et liste les cycles d'une tontine en fonction de sa période.",
                "operationId": "gestionCycle",
                "parameters": [
                    {
                        "name": "tontine",
                        "in": "path",
                        "description": "ID de la tontine",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès - Les cycles de la tontine ont été gérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "status_message": {
                                            "type": "string",
                                            "example": "les cycles du tontine"
                                        },
                                        "cycles": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "dates": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "2024-02-10"
                                            }
                                        },
                                        "nombre_participants": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - L'utilisateur n'est pas le créateur de la tontine"
                    },
                    "400": {
                        "description": "Requête incorrecte - La tontine n'est pas dans un état accepté ou a déjà des cycles"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}